Replace old blockstore with smart store
Old blockstore is replaced with smartstore and smartstore is renamed to blockstore
