+++ code/go/0chain.net/sharder/blockstore/cache.go
@@ -1,5 +1,5 @@
 //cache
-package smartblockstore
+package blockstore
 
 import (
 	"fmt"
