+++ code/go/0chain.net/sharder/blockstore/common.go
@@ -1,4 +1,4 @@
-package smartblockstore
+package blockstore
 
 import (
 	"bufio"
